ported over 0.7.1 and 0.7.1a changes, updated version info

+++ core/src/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/TippedDart.java
@@ -26,6 +26,7 @@ import com.shatteredpixel.shatteredpixeldungeon.ShatteredPixelDungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.PinCushion;
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroSubClass;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.Recipe;
@@ -49,31 +50,42 @@ import java.util.HashMap;
 public abstract class TippedDart extends Dart {
 	
 	{
-		bones = true;
+		tier = 2;
+		
+		//so that slightly more than 1.5x durability is needed for 2 uses
+		baseUses = 0.65f;
 	}
 	
-	@Override
-	public int STRReq(int lvl) {
-		return 11;
-	}
+	//exact same damage as regular darts, despite being higher tier.
 	
 	@Override
 	protected void rangedHit(Char enemy, int cell) {
-		//attempt to stick the dart to the enemy, just drop it if we can't.
-		Dart d = new Dart();
-		if (enemy.isAlive() && sticky) {
-			PinCushion p = Buff.affect(enemy, PinCushion.class);
-			if (p.target == enemy){
-				p.stick(d);
-				return;
+		super.rangedHit( enemy, cell);
+		
+		//need to spawn a dart
+		if (durability <= 0){
+			//attempt to stick the dart to the enemy, just drop it if we can't.
+			Dart d = new Dart();
+			if (enemy.isAlive() && sticky) {
+				PinCushion p = Buff.affect(enemy, PinCushion.class);
+				if (p.target == enemy){
+					p.stick(d);
+					return;
+				}
 			}
+			Dungeon.level.drop( d, enemy.pos ).sprite.drop();
 		}
-		Dungeon.level.drop( d, enemy.pos ).sprite.drop();
 	}
 	
 	@Override
-	public int price() {
-		return 6 * quantity;
+	protected float durabilityPerUse() {
+		float use = super.durabilityPerUse();
+		
+		if (Dungeon.hero.subClass == HeroSubClass.WARDEN){
+			use /= 2f;
+		}
+		
+		return use;
 	}
 	
 	private static HashMap<Class<?extends Plant.Seed>, Class<?extends TippedDart>> types = new HashMap<>();
